rename agenda -> perspectives
